+++ operatorsDelta.h
@@ -5,7 +5,7 @@
 
 
 struct DeltaModularityHash : public thrust::unary_function <thrust::tuple<unsigned long long, float>, float> {
-	float* community_weight;
+	double* community_weight;
 	float* nodes_weight;
 	float total_weight;
 	float* self_community;
@@ -21,13 +21,13 @@ struct DeltaModularityHash : public thrust::unary_function <thrust::tuple<unsign
 		return value;
 	};
 
-	DeltaModularityHash(float* w, float* n, double m, float* s, unsigned int* c) :
+	DeltaModularityHash(double* w, float* n, double m, float* s, unsigned int* c) :
 		community_weight(w), nodes_weight(n), total_weight(m), self_community(s), communities(c) {};
 };
 
 
 struct DeltaModularitySort : public thrust::unary_function <thrust::tuple<unsigned int, unsigned int, float>, float> {
-	float* community_weight;
+	double* community_weight;
 	float* nodes_weight;
 	float total_weight;
 	float* self_community;
@@ -40,7 +40,7 @@ struct DeltaModularitySort : public thrust::unary_function <thrust::tuple<unsign
 		return value;
 	};
 
-	DeltaModularitySort(float* w, float* n, double m, float* s, unsigned int* c) :
+	DeltaModularitySort(double* w, float* n, double m, float* s, unsigned int* c) :
 		community_weight(w), nodes_weight(n), total_weight(m), self_community(s), communities(c) {};
 };
